quick fix

+++ main.py
@@ -2,18 +2,25 @@ import requests
 import zipfile
 import gzip
 import pandas as pd
+import os
 
+# change path 
+file_path = "./draft.csv"
 url = "https://17lands-public.s3.amazonaws.com/analysis_data/draft_data/draft_data_public.FDN.PremierDraft.csv.gz"
 
 response = requests.get(url)
 
-with open("draft_data_public.FDN.PremierDraft.csv.gz", mode="wb") as file:
-    file.write(response.content)
+if not os.path.exists(file_path):
+    print("Fetching data")
+    with open("draft_data_public.FDN.PremierDraft.csv.gz", mode="wb") as file:
+        file.write(response.content)
 
-with gzip.open('draft_data_public.FDN.PremierDraft.csv.gz', 'rb') as f:
-    file_content = f.read()
-    with open('draft.csv', 'wb') as f_out:
-        f_out.write(file_content)
+    with gzip.open('draft_data_public.FDN.PremierDraft.csv.gz', 'rb') as f:
+        file_content = f.read()
+        with open('draft.csv', 'wb') as f_out:
+            f_out.write(file_content)
+else:
+    print("draft.csv already exists")
 
 df = pd.read_csv('draft.csv', nrows=100000)
